requirements.txt
os #could not install
slack_bolt
slack_sdk #already installed from slack_bolt
slack_sdk.errors #could not install
pandas
json #could not install
jsbeautifier
re #could not install
google-api-python-client
slacktastick